more fixes for flatmesh + fix gcc warning + add conversion function for boost.python + switch build off by default due to travis build failure for Py3

+++ src/Mod/MeshPart/App/MeshFlatteningLscmRelax.cpp
@@ -455,7 +455,7 @@ void LscmRelax::lscm()
     std::vector<trip> mat_triplets;
     for (auto triplet: triple_list)
     {
-        if (triplet.col() > static_cast<size_t>((this->vertices.cols() - this->fixed_pins.size()) * 2 - 1))
+        if (triplet.col() > static_cast<int>((this->vertices.cols() - this->fixed_pins.size()) * 2 - 1))
             rhs_triplets.push_back(triplet);
         else
             mat_triplets.push_back(triplet);
